feat:slideAvatar,fix saveImage
