+++ app/src/main/java/be/hogent/faith/faith/drawEmotionAvatar/DrawEmotionViewModel.kt
@@ -1,39 +1,20 @@
 package be.hogent.faith.faith.drawEmotionAvatar
 
-import android.graphics.Bitmap
 import android.graphics.Color
-import android.util.Log
 import androidx.annotation.ColorInt
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import be.hogent.faith.domain.models.Event
 import be.hogent.faith.faith.util.SingleLiveEvent
-import be.hogent.faith.service.usecases.SaveEmotionAvatarUseCase
-import be.hogent.faith.util.TAG
 import com.divyanshu.draw.widget.DrawView
 import com.divyanshu.draw.widget.MyPath
 import com.divyanshu.draw.widget.PaintOptions
-import io.reactivex.disposables.CompositeDisposable
 
 /**
  * ViewModel for the [DrawEmotionAvatarFragment].
  * It mainly holds the state of the [DrawView].
  */
-class DrawEmotionViewModel(private val saveEmotionAvatarUseCase: SaveEmotionAvatarUseCase, private val event: Event) : ViewModel() {
-
-    private val disposables = CompositeDisposable()
-
-    /**
-     * The errormessages
-     */
-    private val _errorMessage = MutableLiveData<String>()
-    val errorMessage: LiveData<String>
-        get() = _errorMessage
-
-    private val _avatarSavedSuccessFully = SingleLiveEvent<Unit>()
-    val avatarSavedSuccessFully: LiveData<Unit>
-        get() = _avatarSavedSuccessFully
+class DrawEmotionViewModel() : ViewModel() {
 
     private val _selectedColor = MutableLiveData<@ColorInt Int>()
     val selectedColor: LiveData<Int>
@@ -87,26 +68,5 @@ class DrawEmotionViewModel(private val saveEmotionAvatarUseCase: SaveEmotionAvat
         _drawnPaths.value = LinkedHashMap()
         _selectedColor.value = Color.BLACK
         _selectedLineWidth.value = LineWidth.MEDIUM
-    }
-
-    override fun onCleared() {
-        disposables.clear()
-        super.onCleared()
-    }
-
-    /**
-     * Save avatar bitmap. This updates the property emotionAvatar
-     */
-    fun saveImage(bitmap: Bitmap) {
-        val saveRequest = saveEmotionAvatarUseCase.execute(
-            SaveEmotionAvatarUseCase.Params(bitmap, event)
-        ).subscribe({
-            Log.d(TAG, "emotionavatar in viewmodel ${event.emotionAvatar?.name?: "no name"}")
-            _avatarSavedSuccessFully.value = Unit
-        }, {
-            _errorMessage.postValue(it.message)
-        }
-        )
-        disposables.add(saveRequest)
     }
 }
